app.py
import os
from flask import Flask, jsonify, request
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

app = Flask(__name__)

# Placeholder for the action routing system
def handle_sync_payouts():
    """Stub function for auto_sync_payouts."""
    return {"status": "stub", "action": "auto_sync_payouts", "message": "Stripe integration function exists, needs API call."}

def handle_market_analysis():
    """Stub function for ai_market_analysis."""
    return {"status": "stub", "action": "ai_market_analysis", "message": "Claude AI analysis function exists, needs API call."}

def handle_powerup():
    """Stub function for infinite_powerup."""
    return {"status": "stub", "action": "infinite_powerup", "message": "Placeholder for infinite_powerup logic."}

# Add more stub functions as needed based on the user's full schema

@app.route('/api/action', methods=['POST'])
def api_action():
    data = request.get_json()
    action_name = data.get('action')
    
    # Simple action router
    if action_name == 'auto_sync_payouts':
        result = handle_sync_payouts()
    elif action_name == 'ai_market_analysis':
        result = handle_market_analysis()
    elif action_name == 'infinite_powerup':
        result = handle_powerup()
    else:
        result = {"status": "error", "message": f"Unknown action: {action_name}"}
        
    # Placeholder for Ledger/audit trail and Service creation tracking
    print(f"Action executed: {action_name}, Result: {result['status']}")
    
    return jsonify(result)

@app.route('/')
def index():
    # This is where the index.html content would be served
    return "SCH Omni System Backend is running. Access the dashboard via index.html."

if __name__ == '__main__':
    port = int(os.environ.get('PORT', 5000))
    app.run(host='0.0.0.0', port=port, debug=os.environ.get('FLASK_ENV') != 'production')
