requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
anthropic==0.18.1
requests==2.31.0
twilio==8.11.0
stripe==7.9.0
